--- a/query/statement_rewriter.go
+++ b/query/statement_rewriter.go
@@ -95,29 +95,16 @@ func rewriteShowMeasurementsStatement(stmt *influxql.ShowMeasurementsStatement) 
 		sources = influxql.Sources{stmt.Source}
 	}
 
-	// Check if we can exclusively use the index via the ShowMeasurementsStatement.
-	if !influxql.HasTimeExpr(stmt.Condition) {
-		stmt.Condition = rewriteSourcesCondition(sources, stmt.Condition)
-		return stmt, nil
+	// Currently time based SHOW MEASUREMENT queries can't be supported because
+	// it's not possible to appropriate set operations such as a negated regex
+	// using the query engine.
+	if influxql.HasTimeExpr(stmt.Condition) {
+		return nil, errors.New("SHOW MEASUREMENTS doesn't support time in WHERE clause")
 	}
 
-	// The query is bounded by time then it will have to query TSM data rather
-	// than utilising the index via system iterators.
-	return &influxql.SelectStatement{
-		Fields: []*influxql.Field{
-			{Expr: &influxql.VarRef{Val: "_name"}, Alias: "name"},
-		},
-		Sources:    rewriteSources2(sources, stmt.Database),
-		Condition:  stmt.Condition,
-		Offset:     stmt.Offset,
-		Limit:      stmt.Limit,
-		SortFields: stmt.SortFields,
-		OmitTime:   true,
-		StripName:  true,
-		EmitName:   "measurements",
-		Dedupe:     true,
-		IsRawQuery: true,
-	}, nil
+	// rewrite condition to push a source measurement into a "_name" tag.
+	stmt.Condition = rewriteSourcesCondition(sources, stmt.Condition)
+	return stmt, nil
 }
 
 func rewriteShowMeasurementCardinalityStatement(stmt *influxql.ShowMeasurementCardinalityStatement) (influxql.Statement, error) {
@@ -384,6 +371,7 @@ func rewriteShowTagKeysStatement(stmt *influxql.ShowTagKeysStatement) (influxql.
 			Alias: "tagKey",
 		},
 	}
+
 	s.Sources = rewriteSources2(stmt.Sources, stmt.Database)
 	return s, nil
 }