###### Required for all non-trivial PRs
- [ ] Rebased/mergable
- [ ] Tests pass

This PR fixes some regressions we had in master, and brings performance of the `SHOW X` commands in line with that of 1.3.7. All `SHOW X` meta queries will now use the index when there is no time component present.

It also: 
 
  - Adds `time` support to `SHOW SERIES` and `SHOW TAG VALUES`
  - Fixes a bug where the `inmem` shared index wouldn't be taken advantage of if one were accessing an `inmem` database from a server using the `tsi1` index by default.
  - Uses the `inmem` optimisation for `SHOW MEASUREMENTS` so that only one shard is queried.

In the case of using `time` with `SHOW MEASUREMENTS` and `SHOW TAG VALUES`, the following should be noted:

  - Currently only `tsi1` is properly supported. If you issue a `SHOW X WHERE time > y` command to a database using the `inmem` index, all matching values across all time will be returned.
  - The `tsi1` index only has time-based support at the shard boundary level. That is to say, we use the time components of the `WHERE` condition to find shards that lie within that range, and return all matching values for those shards.

#### Benchmarks

| Query | v1.3.6 | master | PR |
| ------| ------ | ------ | -- |
| show measurements | 0.030s|  14.930s | 0.030s |
| show measurements with measurement =~ /m./ | 0.031s|  11.954s | 0.026s |
| show measurements where "tag0" = 'value0' | 0.031s|  12.456s | 0.033s |
| show measurements limit 1 | 0.021s|  9.131s | 0.029s |
| show measurements offset 2 | 0.021s|  9.049s | 0.045s |
| show series | 6.920s|  26.776s | 5.793s |
| show series from m0 | 6.015s|  28.463s | 5.150s |
| show series where "tag0" = 'value0' | 6.454s|  28.454s | 5.748s |
| show series limit 1 | 3.939s|  8.918s | 2.783s |
| show series offset 2 | 6.653s|  9.824s | 5.296s |
| show tag keys | 0.014s|  12.291s | 0.014s |
| show tag keys from m0 | 0.014s|  12.877s | 0.015s |
| show tag keys where "tag0" = 'value0' | 0.017s|  12.101s | 0.018s |
| show tag keys limit 1 | 0.018s|  12.312s | 0.024s |
| show tag keys offset 2 | 0.018s|  13.531s | 0.017s |
| show tag values with key = "tag1" | 3.009s|  3.165s | 3.693s |
| show tag values from m0 with key = "tag1" | 3.368s|  3.117s | 3.211s |
| show tag values with key = "tag1" limit 1 | 0.348s|  0.521s | 0.932s |
| show tag values with key = "tag1" offset 2 | 3.864s|  3.114s | 3.260s |
| show measurement exact cardinality | n/a   |  10.583s | 9.163s |
| show measurement exact cardinality group by "tag1" | n/a   |  74.588s | 54.545s |
| show measurement exact cardinality where "tag0" = value0 | n/a   |  0.025s | 0.021s |
| show series exact cardinality | n/a   |  13.039s | 10.290s |
| show series exact cardinality group by "tag1" | n/a   |  28.293s | 62.268s |
| show series exact cardinality where "tag0" = value0 | n/a   |  0.025s | 0.023s |
| show tag key exact cardinality | n/a   |  15.287s | 7.223s |
| show tag key exact cardinality group by "tag1" | n/a   |  43.311s | 62.133s |
| show tag key exact cardinality where "tag0" = value0 | n/a   |  0.027s | 0.021s |
| show tag values exact cardinality with key = "tag1" | n/a   |  19.262s | 12.042s |