@e-dard 2017-11-07
@jwilder @benbjohnson during the CI runs for this PR I have discovered two unrelated races. I have fixed one in 07c4fdc, the other is fixed in 194c4ad.

I cannot reproduce the race, but Appveyor showed it up. From looking at the stack trace it appeared to be a race between `Store.Close` and `Store.monitorShards`. The race should be fixed, but since I can't repro it can you both walk through it and concur you're happy with the fix. The commit message in 194c4ad gives more details.